requirements.txt
langgraph>=0.0.15
langchain-openai>=0.1.0
langchain>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0